Replaced SHA hashing by scrypt
Do not merge this into master, as there are test failures. This commit is mainly to show how an integration of scrypt could be done on the current code base. This requires a newer version of bouncycastle, which is included on this changeset, but Wildfly/EAP already provides this. So, just like one needs to upgrade resteasy before using keycloak, one needs also to update bouncycastle to the version specified on these pom files.

Ignoring the test failures which are not related to SCrypt itself, the feature works as expected. The failures are most likely related to the upgrade of bouncycastle.