KEYCLOAK-2068 - Fix Potential NPE when using Servlet-Filter Adapter.
When using the {{org.keycloak.adapters.servlet.KeycloakOIDCFilter}} a {{NullPointerException}} can be thrown
in the {{org.keycloak.adapters.servlet.FilterSessionStore}} within the {{getParam}} method of the  generated wrapper
in {{buildWrapper}} when the {{content-type}} is not set.
Since the {{content-type}} is only used to parse the body. We simply just check whether the {{body}} is {{null}}
and if so avoid touching the {{content-type}} which prevents the NPE. If the {{body}} is null we return an empty
{{MultivaluedHashMap}} for the parameters.