KEYCLOAK-3348 Introduce updatedTimestamp property to UserEntity
This helps to detect which version of two UserEntity instances is newer.
Propagated updatedTimestamp property through UserModel down to
all UserAdapter and UserEntity implementations.
In case of a JPA based user store the updatedTimestamp
is maintained by a @PrePersist hook in UserEntity.
In case of MongoDB the updatedTimestamp is maintained in
maintained by the Mongo UserAdapter#updateUser.

Signed-off-by: Thomas Darimont <thomas.darimont@gmail.com>

Signed-off-by: Thomas Darimont <thomas.darimont@gmail.com>