KEYCLOAK-3360 Configure inclusion of IDToken on AccessToken refresh
We now allow to configure the inclusion of the IDToken
when refreshing an AccessToken via the `Tokens Tab` in
the Admin-Console.
For backwards compatibility we include the IDToken by default in the
response when requesting an access token refresh.

Previously IDTokens were always generated on access token refresh, but
the OpenID Connect spec states that an IDToken in the
response is optional (see 12.2. Successful Refresh Response in OIDC Spec)
and can therefore be left out.
This spares some computing cycles since no IDToken needs to be generated.