KEYCLOAK-3410 Ease creating user with initial roles via REST
Previously creating a new user with a set of initial realm
and client roles involved sending multiple requests.

We now honor the provided realmRoles and clientRoles
information in the UserRepresentation sent to the
UserResource#create endpoint. This reduces the number of requests
needed to create a user with inital roles to 1.

For symmetry reasons we also return realmRole and clientRole
information via the UserResource#getUser(..) endpoint.

We also reduced the visibility of the UsersResource#updateUserFromRep to
private and removed changed it to be an instance method.
Wasn't called anywhere in the codebase - and shouldn't be called by user
code anyways.

Signed-off-by: Thomas Darimont <thomas.darimont@gmail.com>