KEYCLOAK-1267 Add dedicated SSO timeouts for Remember-Me
Previosuly remember-me sessions where tied to the SSO max session
timeout which could lead to unexpected early session timeouts.
We now allow SSO timeouts to be configured separately for sessions
with enabled remember-me.
This enables users to opt-in for longer session timeouts.

SSO Session Timeouts for RememberMe can now be configured in the
tokens tab in the realm admin console. Max SSO timeouts for Remember-me
need to be equal or greater than normal max SSO timeouts.

Note that most of the work for this PR was done by @cfsnyder and
is based on his earlier PR https://github.com/keycloak/keycloak/pull/2338.
I just adapted his code to the latest Keycloak API changes.

Signed-off-by: Thomas Darimont <thomas.darimont@gmail.com>