KEYCLOAK-3879 Allow search by userId in UserResource#searchForUser
In practice it is often helpful to lookup a user by it's userId.
This PR adds an additional userId lookup when user searches
are performed with a single search-term.
We now try to find a user with it's userId equal to the given searchTerm.
This enables user lookups by their id's in the admin-console via the
user search.