KEYCLOAK-2343 Add support for exact user search
This PoC adds support for using exact matching for
user search operations. Users can use a (currently ugly) checkbox
to mark a search as exact (if checked) or inexact (if unchecked, default).
Previously only inexact (wildcard) search was supported for the
JPA and MongoDB based search.

On the server-side UsersResource I used Java 8 default methods to provide some bridging
for clients that are using the old API.
Note that I couldn't use that approach on the admin-client side
UsersResource since the admin-client module is not (yet?) using Java 8.

Note that the LDAP search only supports exact searches.

Automatic tests are missing but a manual test showed that
everything works as expected. Will add tests if this approach
gets a "GO".