KEYCLOAK-4950 - Include license data from RH-SSO 7.1.0 release.
licenses.xml was manually collected with the assistance of some external scripts. The download-license-files.sh script is included for updating the local license copies when licenses.xml is changed.

maven-resources-plugin must not be used for the individual license files, as it doesn't preserve symlinks. They are included directly by maven-assembly-plugin, which does preserve symlinks.

Properties like ${product.slot} and ${product.name.full} are not available in this branch, but should be used instead of the hardcoded values in pom.xml and assembly.xml when porting to master.