KEYCLOAK-7454 Set product.rhsso.version to 7.3-CD1 (#5229)
This version format may change in the future, but this was the comprimise
agreed on for the purposes of the current tech-preview product build.

This product version string is only used in the UI, logs, etc. So we don't have
to make it OSGI compliant, or anything else that it being used programmatically
would imply.