PoC for PKCE support inf Keycloak JS Adapter
This add support for PKCE with modes `plain` and `S256`.
For support of the `S256`, the base64-js and sha256.js libraries
were included in the keycloak.js file.
The sha256 library is required because the `crypto.subtle.digest`
support is not available for all browsers.
The base64-js library is required because the `btoa` base64 encodings
cannot deal with Uint8ArrayBuffer's and unicode.

This PoC is testes against the latest Keycloak 4.0.0.beta3 but
should also work with older releases that have PKCE support.

To use the PKCE with the JS adapter, just provide a `pkceMethod`
property in the initOptions for the init(..) callback.

An example might look like this:
```
  keycloak.init({
    onLoad: 'login-required',
    checkLoginIframe: true,
    checkLoginIframeInterval: 1,
    pkceMethod: "S256"
  })
```