Update before.adoc
I was confused when I saw that wildfly also ran as part of keycloak and that I was able to deploy to it.  Then afterwards I realised that I should have downloaded a separate instance, but then I was stuck with two vanillas, one on the keycloak wildfly and one on the downloaded wildfly.  Putting in an acknowledgement of the existence of wildfly on the keycloak installation might help alleviate confusion.