KEYCLOAK-9574 Fix broken Role Selection for LDAP Mapper in Admin-Console
We now wrap the result of a role-selection in an JS array and
use the config value as is for rendering, instead of extracting
the first component (which was the first char, since the value
is a string).

Previously any mapper that used role selection could not be
used due to `LDAPMapperCreateCtrl` generating invalid config structures.

Since the component configuration is represented via
org.keycloak.representations.idm.ComponentRepresentation whose
`config` property is a `MultivaluedHashMap`
config values needs to be passed as an Array. However the
LDAPMapperCreateCtrl in (users.js) only passed the role as a String.

Signed-off-by: Thomas Darimont <thomas.darimont@googlemail.com>