params of request-uri may be redirected to keycloak-server to retrieve code
When login-process is outsourced application is called by externel token. When you write your own Authenticator for that kind of token, the token needs to redirected to keycloak-server. This change consists of one additional url-param: kc_redirect_param_name
kc_redirect_param_name stands for the value of the parameter which is to be redirected.