certain params of request-uri( in my case a token from other identity provider ) may be redirected to keycloak-server to login user and retriev a code.
When the login-process is outsourced, the client is called by an externel token, which stands for the loggedin user. When you write your own Authenticator for that kind of token, the token needs to be redirected to the keycloak-server. This change consists of one additional url-param: kc_redirect_param_name 
kc_redirect_param_name stands for the value of the parameter which is to be redirected.
Example: client is called by url {...]&kc_redirect_param_name=token&token=12345yxc
Redirect-URL is [keycloak-server:port]...&token=12345yxc

In keycloak you write your own Authenticator, which can analyze the token and login the user.