When switching to PLAIN in PKCE look for both allowed values
Instead of checking for S256 in code_challenge_method and switching to PLAIN if S256 was not found, check for S256 then PLAIN and report to console as WARNING when none of the allowed values has been found. This helps detecting client programmer errors when they don't set the code_challenge_method properly, think they are using S256 and get PKCE errors because Keycloak switched to PLAIN and only report this when DEBUG level is set (which we usually do not do in production of very busy servers).