KEYCLOAK-9551 Make refresh_token generation for client_credentials optional
Previously Keycloak always generated a refresh_token when client_credentials
grant was used. However, the OAuth 2.0 RFC6749 states that a refresh
token should not be generated when the client_credentials grant is used.
See:
https://tools.ietf.org/html/rfc6749#section-4.4.3

This PR introduces a new in the client setting in the
`OpenID Connect Compatibility Modes` section called
`Skip creation of refresh_token for client_credentials grant`.
If this is `on` refresh_tokens are not generated for client_credential
grant token requests. As a side effect this also destroys the
user-session which is created for the token request of the
service-account user.
The returned token response contains only an access_token and an
optional id_token. A refresh token will not be generated.
If this is `off`, which is also the default, a refresh_token and session
is created as before.