KEYCLOAK-9551 Make refresh_token generation for client_credentials optional
Previously Keycloak always generated a refresh_token when client_credentials
grant was used. However, the [OAuth 2.0 RFC6749](https://tools.ietf.org/html/rfc6749#section-4.4.3) states that a refresh
token should not be generated when the client_credentials grant is used.

This PR introduces a new in the client setting in the `OpenID Connect Compatibility Modes`
section called `Use Refresh Tokens For Client Credentials Grant`.
If this is `on`, then a refresh_token with a valid user-session is generated
for token requests with client_credential grant. This is the default for clients.

If this is `off`, no refresh_token is generated and the user-session
of the service-account user is immediately destroyed after token generation.
The returned `AccessTokenResponse` response contains only an access_token
and an optional id_token. The refresh_token field will be null.

The setting "Use Refresh Tokens For Client Credentials Grant" must be configured explicitly.

Existing clients will be treated as if "use refresh tokens" is enabled to remain backwards compatible.