Improve JS Adapter Documentation Admin Labels
The documentation currently does not use the same wording as the administrative UI. This makes finding the
correct setting more difficult. Additionally, the documentation does not surround the field names in tick
marks as it does in other places. This makes it less obvious what configuration needs to be done.

This commit adjusts the wording to match the administrative UI. It also ensures the field names are enclosed in tick marks.