Add Associazione Rousseau to the adopters
Associazione Rousseau is the non profit that handles the e-voting platform of the main Italian party (5 Stars MoVement)

Our platform uses Keycloak SSO in a Kubernetes cluster, hosted in AWS.

On September the 3rd, we set the worlds record of online voters on a single day (~80k). All was powered by keycloak.

https://is.gd/WcjaYD