Insert nonce from request query string in access token
Problem description
===================

To avoid replication attacks, it is important to handle the nonce
attribute while validating the Access Token. Keycloak's token endpoint
does not use the nonce attribute that the SP has provided in the
request's query string. The token endpoint just uses the nonce value
that is in the user's session.

When using stateless applications I do not have a session to handle it,
but I still would like to have the nonce attribute inside my access
token for the SP.

Proposal
=======

While generating the access token, Keycloak only uses the nonce
attribute of the user's session. I propose to verify if there is a nonce
value in the request's parameters, if it exists, then we override the
session's nonce with the request's nonce while generating the token.