[KEYCLOAK-12713] Fix default redirect URI for broker client in new realm
Problem: The redirect URIs field is not set for the broker client, created
at new realm creation. It is a mandatory field, so it should be populated,
with "none" value.

Cause: The setupBrokerService() method does not populate the redirect URIs
field

Solution: Modify the RealmManager.java:setupBrokerService() method to add
the "none" default redirect URI value