KEYCLOAK-13842 Use correct content-type for requests with XML payload in Admin-Console
We now use an heuristic to detect the use of xml in the
request body and use the content-type application/xml in that case.
In order to apply the logic at all places where data is sent to the
server we use an Angular JS HTTP request interceptor.

Previously the content-type of a request to the client-description-converter
endpoint was always set to application/json.