KEYCLOAK-15532: Add support for File type fields in Authenticator Config
The template for displaying the configuration of authenticators does not currently support the input type of File.

It would be handy to allow files to be uploaded for the authenticator we are developing rather than pasting the contents of the file into a text field.

Template: themes/src/main/resources/theme/base/admin/resources/templates/kc-provider-config.html