KEYCLOAK-17327 - always use KeycloakUriBuilder
The `redirect_url` query parameter, that was generated during the
authentication redirect and the form body parameter generated when
trying to acquire an access token were handled slightly differently:

When redirecting to the authentication server the KeycloakUriBuilder
was not used, when the remote host was a local address. When trying to
acquire an access token later on, it was used in any case. This leads
in our setup to differing urls (https://host:443/... vs https://host/...)
and consequently to an error and no token being handed out.

This commit targets the redirecting behaviour and rearranges the
creation and usage of the KeycloakUriBuilder in so that it conforms to
the behaviour as seen when acquiring an access token.