Change how WebAuthn Credential signature counter is updated
The WebAuthn says that authenticator "SHOULD" implement a signature counter, to aid in detection of cloned authenticators. But this is not a MUST. In the case of some authenticators, e.g. iOS FaceID/TouchID, this signature counter is always zero, which is allowed by the specification. If Keycloak's stored signature counter is incremented by one, such authenticators will not function properly. In this proposed change, the signature counter is updated to the last value of the signature counter provided by the authenticator, which makes it possible to function with zero-signature-counter authenticator, while still being fully compatible with the specification. 

Reference: https://www.w3.org/TR/webauthn-2/#sctn-sign-counter