KEYCLOAK-16863 Compute additional hashes for Github releases tab
A selection of hashes are calculated for every file that will be uploaded
to Github, and the hashes are written into two files that will also be
uploaded: checksums.csv and checksums.xml

The individual .md5 and .sha1 files are not removed by this, but could
be elmininated for brevity if desired, since the csv and xml also
contain these algorithms.

It takes approximately 2 minutes to compute the hashes on my laptop.