KEYCLOAK-10663 Introducing Resource Type in scope based permissions.
Resource based and Scope Based permissions are not merged in single UI because Resource based permission requires resource as compulsory field.

In case of Scope based permission, if Resource Type switch is on, Resource Type field is available and it is compulsory to be filled.
If Resource Type switch is off, it is optional for user to fill Resource field.