KEYCLOAK-19894 Remove SQL lowercase transformation on username db-stored value
The values in the 'username' column of the 'user_entity' table are stored as lower-case values.

When querying that column, the column value in a 'where' condition (unlike the provided search value) needs not be transformed to lower-case again (as it already is). With a transformation in place, indices cannot be used, which hurts performance.

This commit removes the to-lower-case transformation of the 'username' column value on the 'user_entity' table.