KEYCLOAK-19894 Add SQL lowercase transformation on username search input
The values in the 'username' column of the 'user_entity' table are stored as lower-case values.

When querying that column, any input that contains anything but lowercase values will fail.

The existing code-base seems to inconsistently transform input values to lower case. This commit adds to-lower-case transformations in places where that did not yet occur.