Fix NameID Format for saml-user-attribute-nameid-mapper
Currently, if created a User Attribute Mapper For NameID protocol mapper and chosen a different NameID Format as the client's NameID Format, the NameID and its format cannot be set by the mapper properly. Specifically, if a NameID Format that Keycloak does not natively support is chosen on the mapper, this protocol mapper does not work.