Ensure single values are not wrapped as an array
When the claim value is expected to be a single value, i.e. the user interface toggle for multi-value is set to off, the user expects a singular value to set in the token.
Currently, the code converts an array to string, which results in [ the_claim_value ], i.e. value is wrapped as an array. 

The purpose of this pull request is to ensure that there is a non-zero length array, in which case the first value is returned, as described by the popup tooltip in the UI, else an empty string is returned, which can only occur if the user has no mapped roles.

This change can be tested by...

1. Create a user and assign multiple realm roles, having the desired target role being the first assigned.
2. Create a mapper which sets a non-multivalued claim based on the user's role. 
3. Then authenticate as the user and check the mapper has created a claim which is a plain string, e.g.   name_of_my_first_role 
4. The test must fail if the value looks like [ name_of_my_first_role ]  
5. Repeat the test for assigning user client role. I am not certain if this pull request will solve for realm roles AND client roles. Happy to apply the same fix if someone can point me at the code, if indeed it is a different handler to this use case.