Added code to allow a client to override the session idle and max lifetime realm values.
Currently the fields in a client for client session idle and client max lifetime in the admin portal take values to override the realm settings, but the values do not get used. The code in this PR will add the check to the AuthenticationManager. The code works with both cookie and api usage paths, and integration testing has been added for the code updates.

Closes #11959