Revise AuthenticationSelectionResolver (#12102)
Previously the order of the generated AuthenticationExecutionOptions
was partially based on the order of user credentials, and  not
on the order of the Authentication flow definition.

This change simplifies the logic of the AuthenticationSelectionResolver and
collects the Authentication options in the order they are defined in the
corresponding Authentication flow.

The new behavior is now easier to understand because the order of the
executions in the authentication flow is now followed exactly when
generating the authentication options.

Note: This might change the order in which authentication options
are presented for some users.

Adapt MultiFactorAuthenticationTest to reflect fixed ordering authenticator ordering (#12102)

The authentication flow for the test
- UsernameFormFactory Required
- SubFlow Required
-- Password Form Alternative
-- OTP Form Alternative

The order of credentials for the "user-with-two-configured-otp" user is:
- otp (first)
- password
- otp (second)

Hence, the test actually tested the broken behavior by expecting the OTPForm after the UsernameForm, because the
authentication options were based on the order of the stored credentials(!) and not in the order
of the authenticators from the authentication flow.

The "correct" order is UsernameForm -> PasswordForm (try another way -> OTP Form), iff the user has a password configured.
The adapted test reflects this new behaviour.

Signed-off-by: Thomas Darimont <thomas.darimont@googlemail.com>