Implements the jwt-bearer grant type as described in RFC 7523 for utilizing JWTs as authorization grants.
This implementation adds an additional Assertion Grant grant type for clients which enables the client to submit signed JWTs to the token endpoint and receive an access token for the specified user. It also adds a configuration block for OIDC clients where administrators can configure trusted issuer configurations. When Keycloak receives an assertion grant request from a client, it will verify the signed JWT against the certificate, issuer, and audience specified in the trusted issuer configs and check the client has the 'impersonate' permission for the specified user. If no configs successfully validate the token or if the client is not allowed to impersonate the specified user, the request fails.

Closes #24509

Signed-off-by: Ben Cresitello-Dittmar <bcresitellodittmar@mitre.org>