Implement functionality for Keycloak to integrate with external attribute stores and implement default providers for querying attributes from a REST API and sending and receiving SAML attribute query requests and responses based on section 6 of the SAML profile (https://docs.oasis-open.org/security/saml/v2.0/saml-profiles-2.0-os.pdf). This implementation allows administrators to configure trusted attribute sources and sync attributes received to the internal user store or use them as claims in generated access tokens.
This functionality is implemented using a new Attribute Store SPI which is responsible for reaching out to the external attribute store and collecting attributes. To utilize the attributes in keycloak, administrators can periodically sync attributes to the internal user store using the implemented User Store provider or directly populate access token claims using the implemented Attribute Store protocol mapper.

Improvements to the UI for custom user storage providers were made to allow for additional customization of the generated configuration interface in order to support the configuration of mappers for custom providers without customizing the configuration UI.

Closes #29276

Signed-off-by: Ben Cresitello-Dittmar <bcresitellodittmar@mitre.org>