Fixes the "Page Not Found - We are Sorry" after successful login and logout from a desktop app using KeycloakInstalled.
Per @stianst comment on PR 12796 "We need to fix KeycloakInstalled instead. The '/delegated' endpoint was a broken concept that should never have been introduced. It also has a CSRF issue as it's an open endpoint anyone can link to that will invalidate a users session. The proper fix here is to remove getRedirectUrl. Instead of redirecting it should show a message that the browser tab can be closed like it did previously".

The implementation here attempts to do just that. Allows customization of the login success and logout success messages with plain text messages or optionally, by specifying the Path to a customized static single HTML page (the reason it must be a single HTML page is that css and images are not served separately because webserver shuts down immediately after serving the single page). Removed unused imports.

Closes keycloak#9401