Fix required action path for URI building
When building the URI to process required actions, the resolution of the path section is done by looking for the @Path of a method with a specific name. The method was renamed and split long ago, which broke this way to resolve it. This commit changes this by building the path the same way it is done for reset password: by appending the actual path as it is defined in a constant and used in the methods that map to it.