Add option to Force enable users created on Active Directory
This adds an option to enable new imported users to Active Directory
right after registration.
The purpose of this is to have user accounts enabled on Active Directory
before the user has created a password.
This is accomplished by:
- creating a random `unicodePwd` attribute
- applying 512 to the `userAccountControl` attribute

The accompanying change to the UI is committed to its repo.

Closes #16849