Update ExecuteActionsActionTokenHandler.java
The user email should be set to verified only if the token contains an email. 
ExecuteActionsActionToken can be used to create links which are conveyed to the users directly rather than sent by email. For example, a link to update the password can be displayed as QR-code, or printed, if you don't store user emails. 
In this case, you use "public ExecuteActionsActionToken(String userId, int absoluteExpirationInSecs, List<String> requiredActions, String redirectUri, String clientId)" to create the token, and no email is present, and also the handle doesn't complain if there is no email, but then it should be marked verified.