Update example custom cache configuration for v>21
If the example custom InfiniSpan configuration is used on Keycloak
version 21 or later, the following error message can be observed in the
server startup logs:

```
The protocol 'VERIFY_SUSPECT' does not exist in the base stack for
operation
```

In later versions of InfiniSpan, the default transport stacks seems to
use "VERIFY_SUSPECT2" instead of "VERIFY_SUSPECT". This change updates
the example configuration.