Add the ability to use wrapper JDBC drivers.
Keycloak has implemented a custom ClassLoader that ignores requests to load anything containing the word "Driver" to prevent eagerly loading all JDBC drivers that ship with Keycloak and then allows Quarkus to lazily load the specified driver itself.

The issue with this is that when using a wrapper driver (e.g., AWS' advanced JDBC driver), that driver has to delegate to an underlying "real" driver (e.g., Postgres, MySQL, etc).  The KeycloakClassLoader ignores anything containing the word "Driver", so the wrapper driver is never loaded or registered with the DriverManager for the wrapper driver to be able to access the real driver.

This change fixes that issue by adding a flag to the KeyCloakLoader that allows disabling lazy class loading of Drivers.  By default the ClassLoader will continue to ignore all requests to load Driver classes to match the current behavior, but this behavior can be changed with a System property.

For more information, see https://github.com/keycloak/keycloak/discussions/19189 and https://github.com/keycloak/keycloak/issues/13205.

Closes #13205