This is to close [KEYCLOAK-19956](https://issues.redhat.com/browse/KEYCLOAK-19956): Conditional login through identity provider
Added new fields in the OIDC settings page to specify an additional authentication check based on token claims:
- one switch to toggle the check on and off: defaults to off
- one field to define the token name (must be in the ID token section): defaults to empty string
- one field to define the expected token value (with regex matching): defaults to empty string
Latest two fields are hidden if the switch is turned off.

Default values are provided in the IdentityProviderModel, which should guarantee proper upgrade of data from previous release.

I have found no existing unit test on this area, so I did not include a new one for this feature.

Updated Server Admin guide to reflect the change.

Tested on local environment and also on OpenShift deployment.