Use cached policy store in Infinispan PolicyAdapter
Use cached policy store instead of using policy store delegate when
getting associated policies. Without this change Keycloak will make
query to database when getting UMA ticket for every single policy
instead of simply using cached value and its not possible to override
this behaviour in any way. This adds very heavy slowdown on
deployments with a lot of policies, for example for us this increased
the time it takes to get UMA ticket from 300ms to almost 8 seconds and
for every request keycloak was making thousands of database queries.
After this change the time is back to normal and after the data are
already cached Keycloak only makes single query to database.

This is regression as this started happening after upgrade to latest
keycloak. After some digging this was changed in this pull request:
https://github.com/keycloak/keycloak/pull/12797

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>