Fix single namespace mode
The CSV for the operator states that single namespace mode works, but the code only ever watches the operator namespace. By removing the configuration and relying on the Java Operator SDK to apply the right configuration, single namespace works and you can run the operator in one namespace and the workload in another single namespace.

Signed-off-by: James Hewitt <james.hewitt@uk.ibm.com>