HardcodedGroupStorageProvider.java * added searchForGroupByNameNoAncestryStream method
GroupDialogPicker.tsx
* switched places of null and filter on line 83:  isSearching ? null : { search: filter },
* made it a non-toplevel group search on line 95:  count = (await adminClient.groups.count({ search: filter, top: true }))

GroupTree.tsx
* made it a non-toplevel group search on line 95: const count = (await adminClient.groups.count({ search, top: true }))

RealmCacheSession.java
* added searchForGroupByNameNoAncestryStream method

JpaRealmProvider.java
* Changed the definition of getGroupsCountByNameContaining method
this method was using searchForGroupByNameStream to count the results. Which is creates a
hierarchy to top level groups resulting in counting the top level groups and in my dataset this was always 1
By only counting the group ids using the name query "getGroupIdsByNameContaining" we are avoiding
Database hits that was used create the group representation objects.
* added searchForGroupByNameNoAncestryStream method

GroupStorageManager.java
* added searchForGroupByNameNoAncestryStream method

MapGroupProvider.java
* Made the same performance improvement in getGroupsCountByNameContaining() method as in JpaRealmProvider
* added searchForGroupByNameNoAncestryStream method

GroupsResource.java
* in listGroups method we are not calling the searchForGroupByNameNoAncestryStream method.
we are applying the fine grained permissions on the groups in the result set, then
converting to single ancestor hierarchy using the newly introduced GroupUtils.toAncestorsLine() method.

GroupLookupProvider.java
* added searchForGroupByNameNoAncestryStream method to the interface

ModelToRepresentation.java
* added searchForGroupByNameNoAncestryStream method

GroupsResource.java
* modified getGroups endpoint method to use searchForGroupByNameNoAncestryStream if
a search term is used. (this needs more tests in different usages.
since this is part of the group API we may not be aware of all the usages.)

GroupUtils.java
* This is where the main logic of single-line-ancestry is implemented:
toAncestorsLine method uses a static class GroupRepresentationExtended based on
GroupRepresentation class to keep track of the parent ids. it also uses a HashMap
to keep track of processed groups which eliminates repeated database hits.
 List<GroupRepresentationExtended> tree keeps the top-level groups with only the subgroups
 linking the search results are populated.

HardcodedGroupStorageProvider.java
* added searchForGroupByNameNoAncestryStream method