Fix issue with web UI of client-scope protocol mappers with custom IDs
The path parameter mapperId is used on the same path to distinguish both the protocol mapper id during creation and the mapper id during modification. The distinction is made by reading the GUID format, which causes the user interface to crash in the case of a custom format for mapper id. I have added '/create' and '/update' to the end of the path to distinguish the two modes without using the GUID regexp anymore.

Closes #23516