The fix solves the problem when integrating on the infinispan jdbc persistent store.
Symptoms of the problem: when saving sessions to ISPN, they should be saved transparently to jdbc persistent storage, but this was not happening. Finally, I found out that the cache is accessed through a decorator that prevents saving to jdbc storage with the flag: Flag.SKIP_CACHE_STORE. The fix consists in the fact that the administrator has the option to run KC even with a parameter that ignores the given flag.

private static final boolean IGNORE_SKIP_CACHE_STORE = Boolean.getBoolean("keycloak.infinispan.ignoreSkipCacheStore");