Do not join offline user sessions with online client sessions
The findUserSessions* names queries have an INNER JOIN.
When offline sessions are used, the client automatically
gets an online sessions with the same ID as well. The
inner join now returns 2 rows instead of one. This results
in an IllegalStateException in the JpaUserSessionPersisterProvider

This patch adds another condition to the queries to
select only those sessions which have the same value in
the offline column in the database.

Closes #26134

Signed-off-by: Martin Krüger <mkrueger@mkru.de>